+++ packages/search/src/modules/search/__tests__/build-document.unit.spec.ts
@@ -1,3 +1,4 @@
+/// <reference types="jest" />
 import { buildProductDocument, buildCategoryDocument, buildCollectionDocument } from '../lib/build-document'
 
 describe('buildProductDocument', () => {